Create repository and set item and merchant to be it's children
