+++ lib/item.rb
@@ -4,6 +4,7 @@ class Item
   attr_accessor :name,
                 :description,
                 :unit_price,
+                :updated_at
   attr_reader :id,
               :created_at,
               :merchant_id,
