+++ lib/merchant_repository.rb
@@ -1,25 +1,9 @@
 require './lib/merchant'
-
-class MerchantRepository
+require './lib/repository'
+class MerchantRepository < Repository
 
   def initialize(csv_array)
-    @csv_array = csv_array
-  end
-
-  def all
-    @csv_array
-  end
-
-  def find_by_id(id)
-    @csv_array.find do |merchant|
-      merchant.id == id
-    end
-  end
-
-  def find_by_name(name)
-    @csv_array.find do |merchant|
-      merchant.name.casecmp == name.case.cmp
-    end
+    super(csv_array)
   end
 
   def find_all_by_name(name)
@@ -28,28 +12,15 @@ class MerchantRepository
     end
   end
 
-  def max_id_plus_one
-    max = @csv_array.max_by do |merchant|
-      merchant.id
-    end
-    new = max.id.to_i + 1
-    new.to_s
-  end
-
   def create(name)
-    Merchant.new(id: max_id_plus_one,
+    Merchant.new(id: max_id_number_new,
                  name: name)
   end
 
   def update(id, name)
-    new = @csv_array.find do |merchant|
-      merchant.id == id
-    end
+    new = find_by_id(id)
 
     new.name = name
   end
 
-  def delete(id)
-    @csv_array.delete(find_by_id(id))
-  end
 end
